Move from utils to columns_view
