+++ circuits/src/rangecheck/columns.rs
@@ -1,8 +1,8 @@
 use itertools::Itertools;
 use plonky2::field::types::Field;
 
+use crate::columns_view::{columns_view_impl, make_col_map, NumberOfColumns};
 use crate::cross_table_lookup::Column;
-use crate::utils::{columns_view_impl, make_col_map, NumberOfColumns};
 
 #[repr(C)]
 #[derive(Clone, Copy, Eq, PartialEq, Debug)]
